Fix: 상세페이지 레이아웃 컴포넌트 통일
