+++ src/layout/DetailLayout/DetailLayout.module.scss
@@ -1,0 +1,5 @@
+.section {
+  @include column-flexbox();
+  margin-top: 12.87rem;
+  gap: 14.1rem;
+}
