+++ src/pages/CourseDetailPage/CourseDetailPage.module.scss
@@ -1,9 +1,3 @@
-.section {
-  @include column-flexbox();
-  margin-top: 12.87rem;
-  gap: 14.1rem;
-}
-
 .article {
   @include inline-flexbox($ai: start);
   gap: 18.9rem;
