+++ src/pages/SpotDetailPage/SpotDetailPage.module.scss
@@ -1,8 +1,3 @@
-.section {
-  @include column-flexbox();
-  margin-top: 12.87rem;
-}
-
 .article {
   @include inline-flexbox();
   gap: 9.4rem;
